pyproject.toml
[project]
name = "llm_calendar_assistant"
version = "0.1.0"
description = "LLM Calendar Assistant"
authors = [{ name = "Thomas Moesl", email = "thomas.moesl@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "instructor>=1.8.2",
    "openai>=1.79.0",
    "pydantic[email]>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "tzlocal>=5.3.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -vv -q"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
